refactor: reorganize folder structure (core and integrations)
Add core folder and update references.
