+++ .github/workflows/docs.yml
@@ -1,0 +1,28 @@
+name: panvimdoc
+
+on:
+  push:
+    branches:
+      - main
+    paths:
+      - 'DOCUMENTATION.md'
+  workflow_dispatch:
+
+jobs:
+  docs:
+    runs-on: ubuntu-latest
+    name: panvimdoc
+    steps:
+      - uses: actions/checkout@v4
+      - name: panvimdoc
+        uses: kdheepak/panvimdoc@main
+        with:
+          vimdoc: swift.nvim
+          version: "Neovim >= 0.9.0"
+          demojify: true
+          treesitter: true
+          pandoc: "DOCUMENTATION.md"
+      - uses: stefanzweifel/git-auto-commit-action@v5
+        with:
+          commit_message: "chore(docs): auto-generate vimdoc"
+          branch: ${{ github.head_ref }}
