+++ lua/swift/core/config.lua
